requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pymysql>=1.0.2
python-dotenv>=0.19.0
alembic>=1.7.1
pydantic>=1.8.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
pydantic-settings>=2.0.0
